+++ src/main/java/org/visuals/legacy/animatium/command/AnimatiumCommand.java
@@ -140,7 +140,6 @@ public class AnimatiumCommand implements Command<FabricClientCommandSource> {
 
         @Override
         public int run(final CommandContext<FabricClientCommandSource> context) {
-            ConfigUtil.put("onboarding", true);
             final Minecraft minecraft = context.getSource().getClient();
             minecraft.schedule(() -> minecraft.gui.setScreen(new OnboardingScreen(minecraft.gui.screen(), true)));
             return Command.SINGLE_SUCCESS;